model/similarity_test/requirements.txt
openai 
faiss-cpu 
pandas
datasets
